Create hydrate Model.
Any instance large as Eloquent Model is big enough already for Livewire to slow down the component lifecycle, which may lead to poor performance on live updates. For example, if you have a component that represents the user profile (email and username), it's better to pass these parameters to properties as strings instead of the assignment of the whole model and extracting its attributes in the view. Alternatively, you can retrieve the model and then keep it as a protected property. You'll still be able to access the protected objects in Blade templates by using $this->yourProperty, but remember that protected properties should be re-hydrated manually on subsequent requests.' give me this kinds of example.